+++ modules/cpu/geometry/Subdivision.cpp
@@ -26,8 +26,18 @@ void Subdivision::commit()
   createEmbreeGeometry(RTC_GEOMETRY_TYPE_SUBDIVISION);
 
   vertexData = getParamDataT<vec3f>("vertex.position", true);
-  colorsData = getParamDataT<vec4f>("vertex.color");
-  texcoordData = getParamDataT<vec2f>("vertex.texcoord");
+  colorsData = getParamDataT<vec4f>("color");
+  bool isColorsFaceVarying = true;
+  if (!colorsData) {
+    colorsData = getParamDataT<vec4f>("vertex.color");
+    isColorsFaceVarying = false;
+  }
+  texcoordData = getParamDataT<vec2f>("texcoord");
+  bool isTexcoordFaceVarying = true;
+  if (!texcoordData) {
+    texcoordData = getParamDataT<vec2f>("vertex.texcoord");
+    isTexcoordFaceVarying = false;
+  }
 
   level = getParam<float>("level", 5.f);
 
@@ -99,16 +109,37 @@ void Subdivision::commit()
     }
   }
 
+  if ((colorsData && isColorsFaceVarying)
+      || (texcoordData && isTexcoordFaceVarying)) {
+    if (generatedIndicesSize != indexData->size()) {
+      auto data = new Data(OSP_UINT, vec3ui(indexData->size(), 1, 1));
+      generatedIndicesData = &(data->as<uint32_t>());
+      data->refDec();
+      uint32_t *dataptr = static_cast<uint32_t *>(generatedIndicesData->data());
+      for (unsigned int i = 0; i < indexData->size(); i++)
+        *(dataptr++) = i;
+      generatedIndicesSize = indexData->size();
+    }
+    rtcSetGeometryTopologyCount(embreeGeometry, 2);
+    setEmbreeGeometryBuffer(
+        embreeGeometry, RTC_BUFFER_TYPE_INDEX, generatedIndicesData, 1);
+    rtcSetGeometrySubdivisionMode(embreeGeometry, 1, (RTCSubdivisionMode)mode);
+  }
+
   if (colorsData) {
     rtcSetGeometryVertexAttributeCount(embreeGeometry, 1);
     setEmbreeGeometryBuffer(
-        embreeGeometry, RTC_BUFFER_TYPE_VERTEX_ATTRIBUTE, colorsData);
+        embreeGeometry, RTC_BUFFER_TYPE_VERTEX_ATTRIBUTE, colorsData, 0);
+    if (isColorsFaceVarying)
+      rtcSetGeometryVertexAttributeTopology(embreeGeometry, 0, 1);
   }
 
   if (texcoordData) {
     rtcSetGeometryVertexAttributeCount(embreeGeometry, 2);
     setEmbreeGeometryBuffer(
         embreeGeometry, RTC_BUFFER_TYPE_VERTEX_ATTRIBUTE, texcoordData, 1);
+    if (isTexcoordFaceVarying)
+      rtcSetGeometryVertexAttributeTopology(embreeGeometry, 1, 1);
   }
 
   if (!indexLevelData)